Book list now persists to local storage

+++ appES6.js
@@ -47,6 +47,44 @@ class UI {
   }
 }
 
+class Store {
+  static getBooks() {
+    let books;
+    if (localStorage.getItem("books") === null) {
+      books = [];
+    } else {
+      books = JSON.parse(localStorage.getItem("books"));
+    }
+
+    return books;
+  }
+
+  static displayBooks() {
+    const books = Store.getBooks();
+
+    books.forEach(book => {
+      const ui = new UI();
+      ui.addBookToList(book);
+    });
+  }
+
+  static addBook(book) {
+    const books = Store.getBooks();
+    books.push(book);
+    localStorage.setItem("books", JSON.stringify(books));
+  }
+
+  static removeBook(isbn) {
+    let books = Store.getBooks();
+    books = books.filter(book => book.isbn !== isbn);
+    localStorage.setItem("books", JSON.stringify(books));
+  }
+}
+
+document.addEventListener("DOMContentLoaded", function() {
+  Store.displayBooks();
+});
+
 document.getElementById("book-form").addEventListener("submit", function(e) {
   e.preventDefault();
 
@@ -63,12 +101,12 @@ document.getElementById("book-form").addEventListener("submit", function(e) {
   } else {
     ui.addBookToList(book);
 
+    Store.addBook(book);
+
     ui.showAlert("Book Added!", "success");
 
     ui.clearFields();
   }
-
-  //   console.log(book);
 });
 
 document.getElementById("book-list").addEventListener("click", function(e) {
@@ -76,5 +114,8 @@ document.getElementById("book-list").addEventListener("click", function(e) {
 
   const ui = new UI();
 
-  ui.deleteBook(e.target, "Book removed", "success");
+  if (e.target.className === "delete") {
+    ui.deleteBook(e.target, "Book removed", "success");
+    Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
+  }
 });
